heroic: load game info for GOG and Epic games
This introduces the new helper script find-heroic-game-installation.sh
to find games installed by Heroic. It does a bit more work than its
sister script for Steam, find_library_for_file.sh; namely, it looks up
the Wine version and prefix folder set for the game. We need these
because we can't use protontricks later.

In load_gameinfo.sh, we now search in both Steam and Heroic locations to
find the selected game. The current logic is short-circuiting, so if a
Steam game is found, Heroic will not be checked. We could consider
checking both and allowing the user to choose which version they want;
however, I think the likelihood of someone installing the same game in
two different launchers is low enough to not justify the extra code.

This commit downgrades an absent appid from an error to a warning. This
allows us to incrementally add support for GOG/Epic games before adding
Steam support.

We set a new variable, game_launcher, which has the value of "steam" or
"heroic", allowing other scripts to implement different logic for each
launcher.