add heroic support to NXM handler
This makes an important change to our instance directories: we are now
using variables.sh instead of appid.txt.

The latter is inextensible -- we can't just add heroic-related
information to it without using error-prone parsing to determine if it
is a new or old version of that file.

The new file variables.sh can be easily sourced by
modorganizer2-nxm-broker.sh, and can be also extended to add more
information over time as needed.

Some extra code has been added for backwards-compatibility when an
appid.txt file is found.