winetricks.sh: use and prefer downloaded winetricks if present
This uses the winetricks copy downloaded in download_external_resources
.sh and installed to ~/.local/share/modorganizer2/winetricks. Curiously,
it seems like this file was completely unused.

I'm happy to put it to use, as the winetricks verbs are more likely to
succeed the newer the version of winetricks we use, and the downloaded
one is the bleeding-edge latest release.