add placeholder URLs for script extenders across supported games
This prevents load_gameinfo.sh from failing when it tries to determine
`game_scriptextender_url` when the store-specific key is missing.

This is a revision of an earlier patch - decided it was better to add
placeholder values.

We could default expansion of
`game_scriptextender_urls` to the empty string, but doing this defeats
the purpose of `set -u` by leaving us unable to detect typos in these
gamesinfo scripts. This way is also more discoverable to the casual
reader how to add a scriptextender to a game.