configure_steam_wineprefix.sh: check error of protontricks.sh script
This script runs with an execution context of:
set -eu
set -o pipefail

Because of this, in the "steam" case, when protontricks.sh exits
unsuccessfully, the whole pipeline it's in shares its exit code
(because of pipefail). This in turn causes the entire installer to exit
(because of set -e).

This makes it impossible to display the "Error while installing
winetricks" dialog, because our installer exits before we have the
chance to test [ "$?" != "0" ]. Instead we show the default "Operation
canceled" dialog from our exit trap.

This patch wraps all the tricks.sh work in a conditional test, which
makes it `set -e` safe and allows us to display the correct dialog.