configure_steam_wineprefix.sh: allow winetricks error to be ignored
There is a possibility that MO2 will work anyway, or the user would
prefer to mess around with running protontricks/winetricks themselves to
resolve the problem. To prevent this issue from becoming a hard block
to their progress, allow them to proceed with the installation by
answering "Yes" to a dangerquestion.