configure_steam_wineprefix.sh: add MO2 winetricks to game_protontricks
This script is intended to skip running winetricks if there are none to
install. However, we must always install arial and fontsmooth=rgb for
MO2 itself. Originally, these were copied into the game_protontricks
list for each game in gamesinfo/, so our test [ -n
"${game_protontricks[*]}" ] would correctly always run them.

However, in commit 8f696323 we removed these and placed them inline in
the script. This had the undesired effect of causing the above test to
fail if the game itself did not have any other tricks to install. This
meant we would not install arial and fontsmooth=rgb.

This patch makes a slight change. Rather than adding these args to the
final apply command, we add them to the game_protontricks list. This
makes our test correctly consider arial and fontsmooth=rgb when deciding
to run or not to run winetricks.

The test has also been changed to examine the number of items in the
list rather than building a string and testing the number of characters
in it, which is a bit more reliable.